pyproject.toml
[project]
name = "admin-proc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adjusttext>=1.3.0",
    "beautifulsoup4>=4.14.2",
    "google-api-python-client>=2.187.0",
    "ipykernel>=7.1.0",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-google-community>=3.0.1",
    "langchain-google-genai>=3.1.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.3",
    "langchain-tavily>=0.2.13",
    "langgraph>=1.0.3",
    "matplotlib>=3.10.7",
    "networkx>=3.4.2",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "tavily-python>=0.7.13",
    "tenacity>=9.1.2",
]

[tool.pyright]
typeCheckingMode = "basic"
include = ["src"]
exclude = ["**/__pycache__"]
pythonVersion = "3.12"
reportUnusedVariable = "none"

[tool.ruff]
src = ["src"]
exclude = [
    ".venv",
    ".git",
    "__pypackages__",
    ".ruff_cache",
]
line-length = 128

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false
